requirements.txt
opencv-python
numpy
pillow
pytesseract
openai
requests
pyttsx3
pyautogui
pyinstaller
flask
gunicorn
werkzeug
psycopg2-binary
wmi
pystray